Weapon gets it's projectile definiton, if it needs it
[git-p4: depot-paths = "//depot/C35/Students/ahadk/SD/Doomenstein/": change = 457826]

+++ Code/Game/Gameplay/Actors/Player.cpp
@@ -113,7 +113,7 @@ void Player::SwitchWeapon( int weaponToSwitchTo )
 	}
 	else if( weaponToSwitchTo <= -1 )
 	{
-		weaponToSwitchTo = actorWeapons.size() - 1;
+		weaponToSwitchTo = static_cast<int>( actorWeapons.size() ) - 1;
 	}
 
 	possessedActor->m_equippedWeapon = actorWeapons[ weaponToSwitchTo ];
